move to pnpm/vite/packages
This restructures the repository into a "monorepo"-ish thing where
`react-focus-rings` is a single package and the examples are a separate
package, allowing the examples to actually function like they would from
a consumer's perspective.

It also switches to using Vite for building rather than
parcel/microbundle to simplify a lot of the process and use more modern
tools.

Finally it uses pnpm for package management now for similar reasons:
speed, modernity, and simplicity with workspaces.